Tokens en login & delete

+++ SpringBoot/SpringBoot/src/main/java/com/Tortuga/SpringBoot/controllers/UserController.java
@@ -1,8 +1,10 @@
 package com.Tortuga.SpringBoot.controllers;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import com.Tortuga.SpringBoot.DAO.UserDAO;
+import com.Tortuga.SpringBoot.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,11 +20,20 @@ public class UserController {
 	@Autowired
 	private UserDAO userDAO;
 
+	@Autowired
+	private JWTUtil jwtUtil;
 
 	//Obtener usuario
 	@RequestMapping(value= "api/users", method = RequestMethod.GET)
-	public List<User> getUsers() {
+	public List<User> getUsers(@RequestHeader (value = "Authorization") String token) {
+		if (!tokenValidation(token)){return null;}
+
 		return userDAO.getUsers();
+	}
+
+	private boolean tokenValidation (String token){
+		String userID = jwtUtil.getKey(token);
+		return userID != null;
 	}
 	@RequestMapping(value= "api/users/register", method = RequestMethod.POST)
 	public void registerUser(@RequestBody User user) {
@@ -33,11 +44,15 @@ public class UserController {
 	}
 	@RequestMapping(value = "api/users/{id}", method = RequestMethod.GET)
 	public User getUser(@PathVariable Integer id) {
-	return userDAO.getUserById(id);
+
+
+		return userDAO.getUserById(id);
 	}
 
 	@RequestMapping(value = "api/users/{id}", method = RequestMethod.DELETE)
-	public void deleteUser(@PathVariable Integer id ) {
+	public void deleteUser(@RequestHeader (value = "Authorization") String token,
+						   @PathVariable Integer id ) {
+		if (!tokenValidation(token)){return;}
 		userDAO.delete(id);
 	}
